Add fade in animation
Basic, for one element, but it is a great proof of concept. 

Also realized that trying to separate out the Minecraft/Hero rendering was too ambitious; it will simply be clean code that is tied together/well documented (it's not happening otherwise, or perhaps that will be a later term refactor).

+++ src/main/java/gr8pefish/heroreactions/hero/client/RenderingUtils.java
@@ -61,7 +61,7 @@ public class RenderingUtils {
      */
     public static void renderReaction(double currentTime, double totalTime, FeedbackTypes feedbackType, double feedbackRatioOfTotal, List<TransformationTypes> transformationTypes) {
         for (TransformationTypes transformationType : transformationTypes) {
-            transformationType.render(currentTime, totalTime, feedbackType, feedbackRatioOfTotal);
+            transformationType.render((long)currentTime, (long)totalTime, feedbackType, feedbackRatioOfTotal);
         }
     }
